% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{adaptive.Kernelsize.Directionality}
\alias{adaptive.Kernelsize.Directionality}
\title{Adaptive Kernel size}
\usage{
adaptive.Kernelsize.Directionality(n = 3)
}
\arguments{
\item{n}{the amount of connecting pixels. kernelsize = 2n-1}
}
\value{
a list containing, vertical, horizontal, and two diagonal kernels
}
\description{
Adaptive Kernel size
}
\examples{
adaptive.Kernelsize.Directionality(n = 5)
}
