% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FillingTheGap.R
\name{analyze_holes}
\alias{analyze_holes}
\title{Analyse hole characteristica}
\usage{
analyze_holes(img, select.layer = 2)
}
\arguments{
\item{img}{Segmented Root Image Input}

\item{select.layer}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `2`.}
}
\value{
a summary statistics
}
\description{
Analyse hole characteristica
}
\examples{
data(seg_Oulanka2023_Session01_T067)
im = terra::rast(seg_Oulanka2023_Session01_T067)
analyze_holes(im)
}
