% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateZoning.R
\name{binning}
\alias{binning}
\title{Takes a continues depth map and bins it to a specified range}
\usage{
binning(depthmap, nn, round.option = "rounding")
}
\arguments{
\item{depthmap}{1-layer raster, takes output from create.depthmap()}

\item{nn}{bin width}

\item{round.option}{choose the binning operation. Available are "rounding", "ceiling", and "floor".}
}
\value{
raster with input depths in bins
}
\description{
Takes a continues depth map and bins it to a specified range
}
\examples{
img = seg_Oulanka2023_Session01_T067[[2]]
mask = seg_Oulanka2023_Session01_T067[[1]] - seg_Oulanka2023_Session01_T067[[2]]
mask[mask == 255] <- NA
depthmap = create.depthmap(img,mask,start.soil = 290 )
binned.map = binning(depthmap,nn = 5)
}
