% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{circular_mean}
\alias{circular_mean}
\title{Calculate a circular mean to determine average Directionality}
\usage{
circular_mean(angles, input_units = "degrees", output_units = "degrees")
}
\arguments{
\item{angles}{input angles}

\item{input_units}{can be either "radians" or "degrees"}

\item{output_units}{can be either "radians" or "degrees"}
}
\value{
average angle
}
\description{
Calculate a circular mean to determine average Directionality
}
\examples{
circular_mean(angles = c(360,90,0), input_units = "degrees", output_units = "degrees")
}
