% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_stitching3.R
\name{combine_images}
\alias{combine_images}
\title{Combine Two Warped Images}
\usage{
combine_images(warped_prev, warped_curr, overlap_start, overlap_end)
}
\arguments{
\item{warped_prev}{A cimg object representing the first warped image.}

\item{warped_curr}{A cimg object representing the second warped image.}

\item{overlap_start}{An integer specifying the starting column index of the overlap region.}

\item{overlap_end}{An integer specifying the ending column index of the overlap region.}
}
\value{
A cimg object representing the blended and combined image.
}
\description{
Combines two warped images by blending their overlapping regions and concatenating the non-overlapping parts.
}
\keyword{internal}
