% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{convert_to_raster}
\alias{convert_to_raster}
\title{Helper function to convert various input formats to terra raster}
\usage{
convert_to_raster(input)
}
\description{
Helper function to convert various input formats to terra raster
}
\keyword{internal}
