% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{convert_to_spatrast}
\alias{convert_to_spatrast}
\title{Helper function to convert various input formats to terra SpatRaster
Helper function to convert various input formats to terra SpatRaster}
\usage{
convert_to_spatrast(input, normalize = TRUE, select.layer = NULL)
}
\arguments{
\item{input}{Any supported image format}

\item{normalize}{Logical, whether to normalize values to 0-1 range if they're in 0-255}

\item{select.layer}{Numeric, which layer to select if input has multiple layers}
}
\description{
Helper function to convert various input formats to terra SpatRaster
Helper function to convert various input formats to terra SpatRaster
}
\keyword{internal}
