% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats3.R
\name{count_pixels}
\alias{count_pixels}
\title{counts all pixels in a segmented image}
\usage{
count_pixels(img)
}
\arguments{
\item{img}{one layer image}
}
\value{
a numeric value
}
\description{
counts all pixels in a segmented image
}
\examples{
data(seg_Oulanka2023_Session01_T067)
img = terra::rast(seg_Oulanka2023_Session01_T067)[[2]]
rootpixel  = count_pixels(img)
}
