% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_CreateDepthmap_function.R
\name{create.depthmap}
\alias{create.depthmap}
\title{Create A Phase Shifted Cosine Depth Map}
\usage{
create.depthmap(
  im,
  mask,
  tube.thicc = 7,
  tilt = 45,
  dpi = 300,
  start.soil = 0,
  center.offset = 0.5
)
}
\arguments{
\item{im}{input image}

\item{mask}{indicating which pixels are foreign objects like tape. The mask will be: mask = im[[1]] - im[[2]] if 'RootDetector' format is used}

\item{tube.thicc}{Diameter of Minirhizotron Tubes}

\item{tilt}{Minirhiztron Tube insertion angle (typically 30-45 degrees)}

\item{dpi}{Image resolution}

\item{start.soil}{indicates soil boundary 0cm. Can be retrieved from 'SoilSurfE()' but in-situ calibration is recommended}

\item{center.offset}{rotational center. Set 0.5 if down facing tube side is perfectly in the middle}
}
\value{
raster image
}
\description{
Create A Phase Shifted Cosine Depth Map
}
\examples{
map = create.depthmap(im,mask,start.soil = 290 )
}
