% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewDirectionality.R
\name{D8_FlowPath}
\alias{D8_FlowPath}
\title{D8 Flow Direction using slope-based kernels}
\usage{
D8_FlowPath(
  skl,
  depthmap,
  diag.weighted = TRUE,
  kernelsize = 2,
  normalized.center = -1,
  depth.positive = TRUE
)
}
\arguments{
\item{skl}{skeletonized SpatRaster Image}

\item{depthmap}{created by create.depthmap()}

\item{diag.weighted}{diagonal pixels are weighted 1/sqrt(2) to reflect increased diagonal pixel distance.}

\item{kernelsize}{argument used for adaptive.Kernelsize.Directionality(). Indicates the amount of neighboring pixels evaluated i.e., kernelsize 3 = 5x5 kernel
Consider readjusting normalized.center when changing kernel size.}

\item{normalized.center}{kernel center value}

\item{depth.positive}{depth values positive sign}
}
\value{
SpatRaster with deepest D8 neighbor-pixel angles:  0, 45, 90, 135, 180, 225, 270, 315
}
\description{
D8 Flow Direction using slope-based kernels
}
\examples{
library(terra)
library(raster)

data(skl_Oulanka2023_Session03_T067)
skl = terra::rast(skl_Oulanka2023_Session03_T067)[[2]]
depthmap = create.depthmap(skl,skl)
D8_FlowPath(skl,depthmap,kernelsize =2,normalized.center=-1)
}
