% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdrive.R
\name{deep.drive}
\alias{deep.drive}
\title{Deep Drive Estimate}
\usage{
deep.drive(DepthMap = NULL, AngleMap = NULL, RootMap = NULL)
}
\arguments{
\item{DepthMap}{SpatRast with Depth information}

\item{AngleMap}{SpatRast with D8 root angles. See e.g., terra::terrain( v= "flowdir")}

\item{RootMap}{If no Angle MAp is supplied, an AngleMap is calculated from SpatRast containing a segmented, presence-abscence root image.}
}
\value{
ratio of root px with an angle == to the steepest depth slope to the sum of all root px
}
\description{
Deep Drive Estimate
}
\examples{
data(skl_Oulanka2023_Session01_T067)
im = terra::rast(skl_Oulanka2023_Session01_T067)
DepthMap = terra::t(create.depthmap(im,center.offset=0,tube.thicc=3.5))

deep.drive(DepthMap = DepthMap, RootMap = im)
}
