% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{DepthLimiter}
\alias{DepthLimiter}
\title{To cut top or bottom parts of a raster}
\usage{
DepthLimiter(im, fixed.depth = 7200, top.cut = T)
}
\arguments{
\item{im}{image input}

\item{fixed.depth}{how much should be retained}

\item{top.cut}{discard the top part and retain the botom?}
}
\value{
a shorter raster
}
\description{
To cut top or bottom parts of a raster
}
\examples{
im2 = DepthLimiter(im = image, fixed.depth = 7200, top.cut = T)
}
