% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_stitching3.R
\name{extract_overlap_region}
\alias{extract_overlap_region}
\title{Extract Overlap Region from an Image}
\usage{
extract_overlap_region(img, overlap_px = 250, side)
}
\arguments{
\item{img}{A cimg object representing the image.}

\item{overlap_px}{An integer specifying the size of the overlap region in pixels. Default is 200.}

\item{side}{A string specifying the side of the image to extract the overlap from. One of "top", "bottom", "left", or "right".}
}
\value{
A cimg object representing the extracted overlap region.
}
\description{
Extracts a region of an image corresponding to a specified overlap along a given side.
}
\keyword{internal}
