% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FillingTheGap.R
\name{fill_holes}
\alias{fill_holes}
\title{Filling the gaps in root pics}
\usage{
fill_holes(img, max_hole_size = Inf, select.layer = 2, plotting = TRUE)
}
\arguments{
\item{img}{input Image}

\item{max_hole_size}{by default Inf, but you probably want to adjust it}

\item{select.layer}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `2`.}

\item{plotting}{should the result be plotted?}
}
\value{
SpatRast input image but with filled holes
}
\description{
Filling the gaps in root pics
}
\examples{
data(seg_Oulanka2023_Session01_T067)
fill_holes(terra::rast(seg_Oulanka2023_Session01_T067),
select.layer = 2, max_hole_size = 200, plotting = FALSE)
}
