% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_split_join_function.R
\name{join_im}
\alias{join_im}
\title{Fuse two Images}
\usage{
join_im(path, dir.out, pattern = "skeleton")
}
\arguments{
\item{path}{Input path}

\item{dir.out}{Output path}

\item{pattern}{only include images with this pattern}
}
\value{
one complete image
}
\description{
this function is intended to to be used after using the root detector
}
\examples{
join_im(path,out.path) = img
}
