% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lines_diagonal_botleft}
\alias{lines_diagonal_botleft}
\title{Example Image with Diagonal Lines (Bottom-Left to Top-Right)}
\format{
A RasterBrick object with dimensions:
  \itemize{
    \item 191 columns (width)
    \item 99 rows (height)
    \item 4 layers (channels)
  }
}
\source{
Images by J.Cunow
}
\usage{
data(lines_diagonal_botleft)
}
\description{
Test dataset containing an image with diagonal lines running from
bottom-left to top-right for directionality analysis.
}
\details{
The image consists of multiple diagonal lines suitable for testing
directionality detection algorithms.
}
\examples{
\donttest{
  data(lines_diagonal_botleft)
  plot(lines_diagonal_botleft)
}
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
