% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lines_horizontal}
\alias{lines_horizontal}
\title{Example Image with Horizontal Lines}
\format{
A RasterBrick object with dimensions:
  \itemize{
    \item 184 columns (width)
    \item 90 rows (height)
    \item 4 layers (channels)
  }
}
\source{
Images by J.Cunow
}
\usage{
data(lines_horizontal)
}
\description{
Test dataset containing an image with approximately horizontal lines for
directionality analysis.
}
\details{
The image consists of multiple horizontal lines suitable for testing
directionality detection algorithms.
}
\examples{
\donttest{
  data(lines_horizontal)
  plot(lines_horizontal)
}
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
