% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lines_vertical}
\alias{lines_vertical}
\title{Example Image with Vertical Lines}
\format{
A RasterBrick object with dimensions:
  \itemize{
    \item 144 columns (width)
    \item 99 rows (height)
    \item 4 layers (channels)
  }
}
\source{
Images by J.Cunow
}
\usage{
data(lines_vertical)
}
\description{
Test dataset containing an image with approximately vertical lines for
directionality analysis.
}
\details{
The image consists of multiple vertical lines suitable for testing
directionality detection algorithms.
}
\examples{
\donttest{
  data(lines_vertical)
  plot(lines_vertical)
}
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
