% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{load_flexible_image}
\alias{load_flexible_image}
\title{Load an image flexibly from file or convert from memory}
\usage{
load_flexible_image(
  input,
  normalize = TRUE,
  output_format = "cimg",
  select.layer = NULL
)
}
\arguments{
\item{input}{File path or image object}

\item{normalize}{Logical, whether to normalize values to 0-1 range if they're in 0-255}

\item{output_format}{Character, one of "cimg", "spatrast", "array", "brick", "raster"}

\item{select.layer}{Numeric, which layer to select if input has multiple layers}
}
\description{
Load an image flexibly from file or convert from memory
}
\keyword{internal}
