% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffCircadian3.R
\name{LRTest}
\alias{LRTest}
\title{Finite sample/ large sample Likelihood ratio test for circadian pattern detection}
\usage{
LRTest(tt, yy, period = 24, FN = TRUE)
}
\arguments{
\item{tt}{Time vector}

\item{yy}{Expression vector}

\item{period}{Period of the since curve. Default is 24.}

\item{FN}{Type of test to use, TRUE = "FN" (finite) or FALSE = "LS" (large samples). Default is finite sample.}
}
\value{
A list of amp, phase, offset, sigma02, sigmaA2, l0, l1, df, stat, and pvalue.
model: \eqn{y= A * sin(2*pi*x+B)+C}
\item{y}{a 1*n vector of data y}
\item{A}{estimated A^hat from fitCurve}
\item{B}{estimated B^hat from fitCurve}
\item{C}{estimated C^hat from fitCurve}
\item{sigma0}{sigma0^hat under H0}
\item{sigmaA}{sigmaA^hat under H1}
\item{n}{length of data y}
\item{df0}{df under H0}
\item{df1}{df under H1}

 \item{amp}{Amplitude based on formula 1}
 \item{phase}{Phase based on formula 1, phase is restricted within (0, period)}
 \item{peakTime}{Phase based on formula 1, peakTime is restricted within (0, period). phase + peakTime = period/4}
 \item{offset}{Basal level(vertical shift) based on formula 1 or on formula 2}
 \item{sigma02}{Variance estimate under the null (intercept only)}
 \item{sigmaA2}{Variance estimate under the alternative (since curve fitting)}
 \item{l0}{Log likelihood under the null (intercept only)}
 \item{l1}{Log likelihood under the alternative (since curve fitting)}
 \item{df}{Degree of freedom for the LR test}
 \item{stat}{LR statistics}
 \item{pvalue}{P-value from the LR test}
 \item{R2}{Pseudo R2 defined as (tss - rss)/tss}
}
\description{
Test the significance of circadian curve fitting using finite sample likelihood ratio test
}
\details{
Formula 1: \eqn{yy = amp \times sin(2\pi/period \times (phase + tt)) + offset}
Formula 2: \eqn{yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset}

sourced from: https://rdrr.io/github/diffCircadian/diffCircadian/ under GPL3 license
author: Zhiguang Huo, Haocheng Ding
Feb. 24, 2023, 9:07 a.m.; Version 0.0.0
journal article: https://doi.org/10.1093/bib/bbab224
tutorial (for circadian data): http://htmlpreview.github.io/?https://github.com/diffCircadian/diffCircadian/blob/master/vignettes/diffCircadian_tutorial.html
}
\examples{
set.seed(32608)
n <- 50
Period <- 24
tt <- runif(n,0,Period)
Amp <- 2
Phase <- 6
Offset <- 3
yy <- Amp * sin(2*pi/Period * (tt + Phase)) + Offset + rnorm(n,0,1)
LRTest(tt, yy)
}
\author{
Caleb (Zhiguang Huo)
}
\keyword{internal}
