% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinirhizotronR-package.R
\docType{package}
\name{MinirhizotronR-package}
\alias{MinirhizotronR}
\alias{MinirhizotronR-package}
\title{MinirhizotronR: Extracting Depth Resolved Metrics from Minirhizotron Root Scans}
\description{
With MinirhizotronR, analyzing root scans from minirhizotron tubes becomes a piece of cake. It enables the user to extract various root metrics and relate them to a depth distribution, or assess rotational biases. It relies on outside handling of Image Segmentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jcunow/MinirhizotronR}
  \item \url{https://jcunow.github.io/MinirhizotronR/}
  \item Report bugs at \url{https://github.com/jcunow/MinirhizotronR/issues}
}

}
\author{
\strong{Maintainer}: Johannes Cunow \email{johannes.cunow@gmail.com} (\href{https://orcid.org/0009-0006-4830-1162}{ORCID})

}
\keyword{internal}
