% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abind_function.R
\name{name d}
\alias{name d}
\alias{abind2}
\title{abind}
\usage{
abind2(
  ...,
  along = N,
  rev.along = NULL,
  new.names = NULL,
  force.array = TRUE,
  make.names = use.anon.names,
  use.anon.names = FALSE,
  use.first.dimnames = FALSE,
  hier.names = FALSE,
  use.dnns = FALSE
)
}
\arguments{
\item{...}{d}

\item{along}{dim}

\item{rev.along}{d}

\item{new.names}{d}

\item{force.array}{d}

\item{make.names}{d}

\item{use.anon.names}{d}

\item{use.first.dimnames}{d}

\item{hier.names}{d}

\item{use.dnns}{d}
}
\value{
merged multidimensional arrayS
}
\description{
abind
}
\examples{
marray = abind2()
}
