% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexible_inputs.R
\name{normalize_array}
\alias{normalize_array}
\title{Normalize or binarize the array or raster}
\usage{
normalize_array(arr, normalize, binarize)
}
\description{
Normalize or binarize the array or raster
}
\keyword{internal}
