% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thinning.R
\name{prepare_image}
\alias{prepare_image}
\title{Prepare Image for Processing (Internal)}
\usage{
prepare_image(img, layer = 2)
}
\arguments{
\item{img}{A matrix, data frame, or `SpatRaster` object representing the image to be processed.}

\item{layer}{Integer indicating the layer to use if `img` is a multi-layer `SpatRaster`. Default is 2.}
}
\value{
A binary matrix with dimensions corresponding to the input image.
}
\description{
This internal function ensures that the input image is in binary format and a matrix structure.
It accepts matrices, data frames, or `SpatRaster` objects and standardizes the format for further processing.
}
\details{
- If `img` is a `SpatRaster`, it will be converted to a matrix. For multi-layer rasters, the specified `layer` is used.
- If `img` is a data frame, it will be converted to a matrix.
- Ensures the matrix contains binary values (`1` for foreground and `0` for background).
- Validates that the matrix contains at least one foreground pixel (`1`).
}
\examples{
\dontrun{
# Example usage
raster <- terra::rast(matrix(c(0, 1, 0, 1), nrow = 2))
binary_image <- prepare_image(raster)
}
}
\keyword{internal}
