% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{resize.image}
\alias{resize.image}
\title{Resize the Image}
\usage{
resize.image(import.path, output.path, height = 2550, width = 2273)
}
\arguments{
\item{import.path}{Input Path}

\item{output.path}{Output Path}

\item{height}{target height}

\item{width}{target width}
}
\value{
a image with new dimensions and resolution
}
\description{
useful if some scanning campaigns have used different dpi
}
\examples{
resize.image(import.path, output.path) = resized.image
}
