% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rgb_Oulanka2023_Session03_T067}
\alias{rgb_Oulanka2023_Session03_T067}
\title{An original Minirhizotron Root Scan from Timepoint 2}
\format{
a RasterBrick with 4900 columns x 1161 rows x 3 layer
}
\source{
Images by J.Cunow
}
\usage{
data(rgb_Oulanka2023_Session03_T067)
}
\description{
This is a blended image of multiple scans stitched together.
The scan originates from a sedge fen in northern Finland in late October.
Columns correspond to tube length and rows to tube rotation
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
