% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers3.R
\name{root.accumulation}
\alias{root.accumulation}
\title{Calculate root accumulation}
\usage{
root.accumulation(x, group, depth, variable, stdrz = "counts")
}
\arguments{
\item{x}{Data frame containing group, depth, and variable columns}

\item{group}{Character vector specifying grouping variable(s)}

\item{depth}{Character string specifying depth column name}

\item{variable}{Character string specifying accumulating values column}

\item{stdrz}{Character string specifying standardization method}
}
\value{
Numeric vector of accumulated values
}
\description{
Calculate root accumulation
}
\examples{
df = data.frame(depth = c(seq(0,80,20),seq(0,80,20)),
               Plot = c(rep("a",5),rep("b",5)), rootpx = c(5,50,20,15,5,10,40,30,10,5) )
accum_root = root.accumulation(df,group = "Plot", depth = "depth", variable = "rootpx")
}
