% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_RootThicc_function.R
\name{root.tickness}
\alias{root.tickness}
\title{Approximate average Root Thickness}
\usage{
root.tickness(kimuralength, rootpx, dpi)
}
\arguments{
\item{kimuralength}{length of roots in image section}

\item{rootpx}{amount of rootpx in the image section}

\item{dpi}{image resolution}
}
\value{
a value in units cm
}
\description{
Approximate average Root Thickness
}
\examples{
root.ticc = root.tickness(kimuralength = 300,rootpx = 9500,300)
}
