% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats.R
\name{RootLength}
\alias{RootLength}
\title{Calculate Root Length using Kimura's Method with optimizations}
\usage{
RootLength(im, unit = "cm", dpi = 300)
}
\arguments{
\item{im}{A skeletonized root image raster}

\item{unit}{Output unit ("px" or "cm")}

\item{dpi}{Image resolution (required when unit = "cm")}
}
\value{
Numeric value representing root length in specified unit
}
\description{
Calculate Root Length using Kimura's Method with optimizations
}
\examples{
data(skl_Oulanka2023_Session01_T067)
img = terra::rast(skl_Oulanka2023_Session01_T067[[2]])
RL = RootLength(im = img,unit = "cm", dpi = 300)
}
