% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats.R
\name{RootScapeMetrics}
\alias{RootScapeMetrics}
\title{RootScapeMetric relies on Landscapemetrics to extract 'Root Scape' Features akin to landscape analysis.}
\usage{
RootScapeMetrics(
  img,
  indexD = NA,
  select.layer = NULL,
  metrics = c("lsm_c_ca", "lsm_l_ent", "lsm_c_pd", "lsm_c_np", "lsm_c_pland",
    "lsm_c_area_mn", "lsm_c_area_cv", "lsm_c_enn_mn", "lsm_c_enn_cv")
)
}
\arguments{
\item{img}{segmented raster  (values = 0,1). Consider whether skeletonized raster is appropriate.}

\item{indexD}{please specify depth. Will only affect the output column = "depth". Useful when used in a loop.}

\item{select.layer}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `2`.}

\item{metrics}{which ,metrics should be calculated from the available ones in 'landscapemetrics::calculate_lsm()'.}
}
\value{
a bunch of metric values
}
\description{
RootScapeMetric relies on Landscapemetrics to extract 'Root Scape' Features akin to landscape analysis.
}
\examples{
data(seg_Oulanka2023_Session01_T067)
img = terra::rast(seg_Oulanka2023_Session01_T067)[[2]]
RootScapeObject  = RootScapeMetrics(img,indexD = 80, metrics = c("lsm_c_ca"))
}
