% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotateRaster.R
\name{RotateRaster}
\alias{RotateRaster}
\title{Rotate simple features for 3D layers
Rotates a simple features layer using a shear matrix transformation on the
\code{geometry} column. This can get nice for visualisation and works with
points, lines and polygons. Sourced from: https://www.mzes.uni-mannheim.de/socialsciencedatalab/article/geospatial-data/}
\usage{
RotateRaster(data, x_add = 0, y_add = 0, rotation.angle = 45)
}
\arguments{
\item{data}{an object of class \code{sf}}

\item{x_add}{integer; x value to move geometry in space}

\item{y_add}{integer; x value to move geometry in space}

\item{rotation.angle}{the angle of rotation}
}
\value{
raster output with rotation
}
\description{
Rotate simple features for 3D layers
Rotates a simple features layer using a shear matrix transformation on the
\code{geometry} column. This can get nice for visualisation and works with
points, lines and polygons. Sourced from: https://www.mzes.uni-mannheim.de/socialsciencedatalab/article/geospatial-data/
}
\examples{
rotated.raster = RotateRaster(raster,rotation.angle = 45)
}
