% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Rotational_function.R
\name{RotationE}
\alias{RotationE}
\title{Estimates Rotation from Tape Coverage}
\usage{
RotationE(im, tape.brightness = 0.66, extra.rows = 100)
}
\arguments{
\item{im}{image.tiff}

\item{tape.brightness}{used for clustering. Tape appears bright e.g., 0.66}

\item{extra.rows}{In case no tape is present. Best leave unchanged - some extra.rows are recommended and will be substracted from the output anyway.}
}
\value{
numeric value corresponding to the center of extruding tape
}
\description{
Estimates Rotation from Tape Coverage
}
\examples{
value = RotationE(img)
}
