% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Rotational_function.R
\name{RotCensor}
\alias{RotCensor}
\title{Rotation Censor}
\usage{
RotCensor(img, center.offset = 0, cut.buffer = 0.02)
}
\arguments{
\item{img}{The image which should be censored}

\item{center.offset}{Rotational shift in rows. Can be retrieved from 'RotShiftDet()', reference image is 0 offset}

\item{cut.buffer}{fixed ratio that will be cut in addition}
}
\value{
a raster
}
\description{
Rotation Censor
}
\examples{
censored.raster = RotCensor(img,center.offset = 120, cut.buffer = 0.02)
}
