% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{RotShiftDet}
\alias{RotShiftDet}
\title{Rotation Correlation}
\usage{
RotShiftDet(
  img1,
  img2,
  cor.type = "phase",
  fixed.depth.pixel = 4000,
  fixed.width = 1500
)
}
\arguments{
\item{img1}{Reference Image with 3 rgb channel}

\item{img2}{Subsequent Image with 3 rgb channel}

\item{cor.type}{Two correlation types available: "ccf" cross correlation, and "phase" phase correlation in frequency domain. See ??imagefx}

\item{fixed.width}{if alignment is necessary, see 'RotCensor()'}

\item{fixed.rotation.pixel}{if alignment is necessary}
}
\value{
numeric value corresponding to rotation in rows. Ensure correct image rotation i.e., rows == rotation
}
\description{
Rotation Correlation
}
\examples{
img1 = seg_Oulanka2023_Session01_T067
img2 = seg_Oulanka2023_Session03_T067
y.lag = RotShiftDet(img1,img2,"phase")
}
