% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{rrwr}
\alias{rrwr}
\title{read rename write wrapper}
\usage{
rrwr(img, pattern, replace, dir, dir.out)
}
\arguments{
\item{img}{image name}

\item{pattern}{pattern}

\item{replace}{replace}

\item{dir}{where to find the image}

\item{dir.out}{where to write the image}
}
\value{
image output
}
\description{
read rename write wrapper
}
\examples{
rrwr(Oulanka2023_T001_L001.tiff,
pattern = "_L001", replace = "", dir = getwd(),dir.out = paste0(getwd(),"/renamed/"))
}
