% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDistributions.R
\name{RWDI}
\alias{RWDI}
\title{Root Weight Depth Index}
\usage{
RWDI(w, roots)
}
\arguments{
\item{w}{weight, typically a vector containing depths}

\item{roots}{root cover}
}
\value{
a value between 0-1
}
\description{
Root Weight Depth Index
}
\examples{
w = seq(5,25,5)
roots = c(0,10,7,3,1)
RWDI(w,roots)
}
