% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seg_Oulanka2023_Session01_T067}
\alias{seg_Oulanka2023_Session01_T067}
\title{A Minirhizotron Root Scan after Segmentation from Timpoint 1}
\format{
a RasterBrick with 4000 columns x 1600 rows x 3 layer
}
\source{
Images by J.Cunow
}
\usage{
data(seg_Oulanka2023_Session01_T067)
}
\description{
The image is derived from a rgb root scan after using "RootDetector" to segment.
Roots are represented as = 1, Background = 0.
The first layer also shows foreign objects such as tape as = 1.
The scan originates from a sedge fen in northern Finland in early June.
Columns correspond to tube length and rows to tube rotation
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
