% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{skeletonize}
\alias{skeletonize}
\title{Skeletonize a segmented image}
\usage{
skeletonize(img, itr = 2)
}
\arguments{
\item{img}{binary image, can be path or magick object}

\item{itr}{iterations of thinning. see magick::image_morphology}
}
\value{
image in as magick object
}
\description{
Skeletonize a segmented image
}
\examples{
img.skeleton = skeletonize("/path/image.png",itr = 2)
}
