% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{skeletonize_function}
\alias{skeletonize_function}
\title{Skeletonize a segmented image}
\usage{
skeletonize_function(img, method = "gonzales")
}
\arguments{
\item{img}{binary iamge}

\item{method}{which approach should be used}
}
\value{
binary image with area turned into linear features
}
\description{
Skeletonize a segmented image
}
\examples{
skeletonize(img,"gonzales") = skeleton_image
}
