% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{SoilSurfE}
\alias{SoilSurfE}
\title{Estimate the position of the soil surface by tape presence}
\usage{
SoilSurfE(
  img,
  search.area = 0.45,
  tape.tresh = 0.33,
  dpi = 300,
  tape.overlap = 0.5,
  tape.brightness = 0.66,
  extra.rows = 100,
  tape.quantile = 0.98
)
}
\arguments{
\item{img}{raster,filename, or array input}

\item{search.area}{ratio of image which is used to look for tape cover. Speeds up computation.}

\item{tape.tresh}{ratio of how much of the tube rotation needs to covered in tape}

\item{dpi}{image resolution}

\item{tape.overlap}{assumes a safety margin on the tape. The soil surface will be shifted by this amount in cm}

\item{tape.brightness}{used for clustering. Tape appears bright e.g., 0.66}

\item{extra.rows}{In case no tape is present. Best leave unchanged - some extra.rows are recommended and will be subtracted from the output anyway.}

\item{tape.quantile}{aligns extra.rows brightness with the tape. The default uses Silver Tape as reference.}
}
\value{
data.frame with tape end and soil surface estimation in rows
}
\description{
Estimate the position of the soil surface by tape presence
}
\examples{
data.frame(soil0, tape.end) = SoilSurfE(im)
}
