% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{split_im}
\alias{split_im}
\title{Split Images}
\usage{
split_im(path, dir.out, pattern = ".tiff", ratio = 0.5)
}
\arguments{
\item{path}{Input path}

\item{dir.out}{Output path}

\item{pattern}{only include images with this pattern}

\item{ratio}{split point 0-1}
}
\value{
two sets of images
}
\description{
Use if too many roots are present for the Root Detector AI image segmentation
}
\examples{
split_im(path, dir.out, ".tiff") = c(top.im, bot.im)
}
