% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexible_inputs.R
\name{supported_formats_string}
\alias{supported_formats_string}
\title{Get list of supported conversion & input formats as a string}
\usage{
supported_formats_string(type = "Input")
}
\description{
Get list of supported conversion & input formats as a string
}
\keyword{internal}
