% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDistributions.R
\name{tail_weight_function}
\alias{tail_weight_function}
\title{Tail weight function (e.g., exponential weight for tail emphasis)}
\usage{
tail_weight_function(
  index = NULL,
  parameter = list(lambda = 0.2, x0 = 5),
  index.spacing = "equal",
  method = "sigmoid",
  baseline.weight = 0,
  inverse = FALSE
)
}
\arguments{
\item{index}{a positive numeric vector containing probability spacing e.g., depth}

\item{parameter}{list with lambda -> shape parameter (0 = constant weighting) & x0 -> curve offset (= inflexion point? )}

\item{index.spacing}{whether index intervals are equally distant i.e., c(1,2,3,4....n), if "equal" than index is c(1,n)}

\item{method}{weighting function along index. Available options are: c("constant", "asymptotic", "linear, "exponential", "sigmoid", "gompertz","step")}

\item{baseline.weight}{minimal weight between 0-1}

\item{inverse}{changes from right tail to left tail if TRUE}
}
\value{
weights between 0-1 along the index
}
\description{
Tail weight function (e.g., exponential weight for tail emphasis)
}
\keyword{internal}
