% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDistributions.R
\name{tail_weighted_js_divergence}
\alias{tail_weighted_js_divergence}
\title{Compute the tail-weighted Jensen-Shannon divergence}
\usage{
tail_weighted_js_divergence(
  P,
  Q,
  parameter = list(lambda = 0.2, x0 = 30),
  method = "constant",
  inverse = FALSE,
  alignPQ = TRUE,
  index = 1:min(c(length(Q), length(P))),
  index.spacing = "equal"
)
}
\arguments{
\item{P}{probability vector 2}

\item{Q}{probability vector 1}

\item{parameter}{list with lambda -> shape parameter (0 = constant weighting) & x0 -> curve offset (= inflexion point )}

\item{method}{weighting function along index. Available options are: c("constant", "asymptotic", "linear, "exponential", "sigmoid", "gompertz","step")}

\item{inverse}{changes from right tail to left tail if TRUE}

\item{alignPQ}{if TRUE, index end values will be cut off in case of unequal length of P & Q so that length of P & Q is equal}

\item{index}{a positive numeric vector containing probability spacing e.g., depth}

\item{index.spacing}{whether index intervals are equally distant i.e., c(1,2,3,4....n), if "equal" than index is c(1,n)}
}
\value{
Jensen-Shannon Divergence - symmetric version of KL
}
\description{
Compute the tail-weighted Jensen-Shannon divergence
}
\examples{
P <- c(0.025,0.05,0.1,0.15, 0.2, 0.3,0.4, 0.5,0.3,0.1)  # Distribution P
Q <- c(0.025,0.05,0.1,0.15, 0.2, 0.3,0.4, 0.5,0.3,0.1)**6  # Distribution Q

# Ensure the distributions are valid (non-negative and sum to 1)
P <- P / sum(P)
Q <- Q / sum(Q)

tail_weighted_js_divergence(P,Q,parameter = list(lambda = 0.2,x0=30))
}
