% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TubeColoration_function.R
\name{Tube.coloration}
\alias{Tube.coloration}
\title{Color Coordinates and Brightness of an image segment}
\usage{
Tube.coloration(img, r = 0.2126, g = 0.7152, b = 0.0722)
}
\arguments{
\item{img}{raster with 3 layers, first is red, second is green, third is blue}

\item{r}{grayscale factor red}

\item{g}{grayscale factor green}

\item{b}{grayscale factor blue}
}
\value{
vector with red chromatic coordinate, green chromatic coordinate, blue chromatic coordinate, mean brightness, mean luminosity
}
\description{
Color Coordinates and Brightness of an image segment
}
\examples{
c(rcc, gcc, bcc, mean.bright, mean.lum)  = Tube.coloration(img)
}
