% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turnover_functions.R
\name{Turnover.TC}
\alias{Turnover.TC}
\title{Global root production and root turnover}
\usage{
Turnover.TC(im.t1, im.t2, method = "kimura", unit = "cm", dpi = 300)
}
\arguments{
\item{im.t1}{brick raster timepoint 1}

\item{im.t2}{brick raster timepoint 2}

\item{method}{choose between "kimura" & "rootpx". Adjust the input image accordingly}

\item{unit}{unit of root length. only applies if method = "kimura"}

\item{dpi}{image resolution. only applies if method = "kimura"}
}
\value{
standing roots at the start, standing roots at the end, root production, % of new roots compared to starting conditions, % of new roots at the second timepoint
}
\description{
Global root production and root turnover
}
\details{
Chosing kimura as method will return root length with default settings from RootLength()
}
\examples{
tunrover.values = TurnoverE(im.t1 = time1, im.t2 = time2, method= "kimura")
}
