% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TurnoverDPC_data}
\alias{TurnoverDPC_data}
\title{Turnover Estimation of Individual Roots from "RootDetector"}
\format{
a RasterBrick with 2550 columns x 2273 rows x 3 layer
}
\source{
Images by J.Cunow
}
\usage{
data(TurnoverDPC_data)
}
\description{
The image is derived from a rgb root scan after using "RootDetector" to segment and compare against a second time point using the root tracking feature.

Roots are represented as max value, Background = 0.
First layer shows root decay (disappeared between time points), foreign objects such as tape, and no-change roots
Second Layer shows new root growth and no-change roots
Third Layer shows no-change-roots
The scan originates from a sedge fen in northern Finland.
Columns correspond to tube length and rows to tube rotation
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
