% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thinning3.R
\name{validate_image_input}
\alias{validate_image_input}
\title{Validate Image Input Parameters}
\usage{
validate_image_input(
  img,
  allow_empty = FALSE,
  min_dim = c(3, 3),
  require_binary = TRUE,
  select.layer = NULL
)
}
\arguments{
\item{img}{The input image to validate}

\item{allow_empty}{Logical, whether to allow empty images}

\item{min_dim}{Minimum required dimensions}

\item{require_binary}{Logical, whether to require binary values}

\item{select.layer}{Layer to validate for multi-layer images}
}
\value{
List containing validated and processed image data
}
\description{
Internal function to validate input parameters for image processing functions
Used for image skeletonizing
}
\keyword{internal}
