% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffCircadian3.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Helper function to validate inputs}
\usage{
validate_inputs(tt, yy, period, parStart = NULL, beta = NULL)
}
\description{
Helper function to validate inputs
}
\keyword{internal}
