% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateZoning.R
\name{zone.fun}
\alias{zone.fun}
\title{Extract a zone of interest based on depth values}
\usage{
zone.fun(rootpic, binned.map, indexD = 0, nn = 5, silent = FALSE)
}
\arguments{
\item{rootpic}{SpatRaster/matrix/array - source image}

\item{binned.map}{SpatRaster/matrix/array - binned depth map}

\item{indexD}{numeric - depth index to extract}

\item{nn}{numeric - bin width used in binning}

\item{silent}{logical - suppress messages}
}
\value{
SpatRaster - extracted zone
}
\description{
Extract a zone of interest based on depth values
}
\examples{
data(seg_Oulanka2023_Session01_T067)
img = terra::rast(seg_Oulanka2023_Session01_T067[[2]])
mask = seg_Oulanka2023_Session01_T067[[1]] - seg_Oulanka2023_Session01_T067[[2]]
mask[mask == 255] <- NA
depthmap = create.depthmap(img,mask,start.soil = 290 )
binned.map = binning(depthmap,nn = 5)
image.zone = zone.fun(img, binned.map, indexD = 0, nn=5, silent = FALSE)
}
