% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateZoning3.R
\name{zone.rotation.fun}
\alias{zone.rotation.fun}
\title{Extract zones based on rotation axis}
\usage{
zone.rotation.fun(
  img,
  k = c(3, 4),
  kk = 5,
  mm = c(2000, 4000),
  select.layer = NULL
)
}
\arguments{
\item{img}{array/matrix/SpatRaster - source image}

\item{k}{numeric vector - which cuts to keep (length 2)}

\item{kk}{numeric - total number of cuts}

\item{mm}{numeric vector - region limits (length 2)}

\item{select.layer}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `NULL`.}
}
\value{
SpatRaster - extracted rotation zone
}
\description{
Extract zones based on rotation axis
}
\examples{
data(seg_Oulanka2023_Session01_T067)
img = terra::rast(seg_Oulanka2023_Session01_T067)
rotationZone = zone.rotation.fun(img, k = c(1,2), kk = 7, mm = c(1500,3000))
}
