% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateZoning.R
\name{binning}
\alias{binning}
\title{Takes a continues depth map and bins it to a specified range}
\usage{
binning(depthmap, nn, round.option = "standard")
}
\arguments{
\item{depthmap}{1-layer raster, takes output from create.depthmap()}

\item{nn}{bin width}

\item{round.option}{choose the binning operation. Available are "rounding", "ceiling", and "floor".}
}
\value{
raster with input depths in bins
}
\description{
Takes a continues depth map and bins it to a specified range
}
\examples{
image = binning(depthmap,nn = 5)
}
