% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Helper_functions.R
\name{binning}
\alias{binning}
\title{Takes a continues depth map and bins it to a specified range}
\usage{
binning(depthmap, nn)
}
\arguments{
\item{depthmap}{image with depth information as illumination}

\item{nn}{bandwidth in cm}
}
\value{
depths in bins
}
\description{
Takes a continues depth map and bins it to a specified range
}
\examples{
image = binning(depthmap,nn = 5)
}
