% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{blur.correction}
\alias{blur.correction}
\title{Threshoold an image to rebinarize a blurred image (e.g., jpeg compression)}
\usage{
blur.correction(img, threshold = 0.4, focus.layer = 2, mask.layer = 1)
}
\arguments{
\item{img}{raster}

\item{threshold}{ratio of max value assigned as 1}

\item{focus.layer}{which layer should be used to capture blurr}

\item{mask.layer}{preserve mask sections (e.g., tape) in one layer}
}
\value{
raster
}
\description{
Threshoold an image to rebinarize a blurred image (e.g., jpeg compression)
}
\examples{
blurred.img = terra::rast(seg_Oulanka2023_Session03_T067)
img = blur.correction(blurred.img, 0.3)
}
