% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers3.R
\name{blur.correction}
\alias{blur.correction}
\title{Threshold an image to rebinarize a blurred image}
\usage{
blur.correction(img, threshold = 0.4, select.layer = 2, mask.layer = 1)
}
\arguments{
\item{img}{SpatRaster object}

\item{threshold}{Numeric value between 0 and 1 for thresholding}

\item{select.layer}{Integer specifying which layer should be used to capture blur}

\item{mask.layer}{Integer specifying which layer preserves mask sections}
}
\value{
SpatRaster object
}
\description{
Threshold an image to rebinarize a blurred image
}
\examples{
blurred.img = terra::rast(seg_Oulanka2023_Session03_T067)
img = blur.correction(blurred.img, 0.3)
}
