% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDepthmap.R
\name{create.depthmap}
\alias{create.depthmap}
\title{Create A Phase-Shifted, Tilt-Amplitude Sine Depth Map}
\usage{
create.depthmap(
  im,
  mask = NULL,
  sinoid = TRUE,
  tube.thicc = 7,
  tilt = 45,
  dpi = 300,
  start.soil = 0,
  center.offset = 0
)
}
\arguments{
\item{im}{input image}

\item{mask}{indicating which pixels are foreign objects like tape. The mask will be: mask = im[[1]] - im[[2]] if 'RootDetector' format is used}

\item{sinoid}{set sinoid = TRUE if true depth should be used. Otherwise, tube diameter is ignored.}

\item{tube.thicc}{Diameter of minirhizotron tubes}

\item{tilt}{Minirhiztron tube insertion angle (typically 30-45 degrees)}

\item{dpi}{Image resolution}

\item{start.soil}{indicates soil boundary 0cm. Can be retrieved from 'SoilSurfE()' but in-situ calibration is recommended}

\item{center.offset}{rotational center. Set 0 if top facing tube side is perfectly in the middle. Set 1 if upfacing tube is located at the edges}
}
\value{
raster image
}
\description{
Create A Phase-Shifted, Tilt-Amplitude Sine Depth Map
}
\examples{
data(seg_Oulanka2023_Session01_T067)
seg_Oulanka2023_Session01_T067 = terra::rast(seg_Oulanka2023_Session01_T067)
img = seg_Oulanka2023_Session01_T067[[2]]
mask = seg_Oulanka2023_Session01_T067[[1]] - seg_Oulanka2023_Session01_T067[[2]]
mask[mask == 255] <- NA
map = create.depthmap(img,mask,start.soil = 290 )
}
