% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdrive.R
\name{deep.drive}
\alias{deep.drive}
\title{Calculate Deep Drive Estimate for Root Systems}
\usage{
deep.drive(
  DepthMap = NULL,
  AngleMap = NULL,
  RootMap = NULL,
  select.layerRM = 2,
  select.layerDM = NULL,
  select.layerAM = NULL
)
}
\arguments{
\item{DepthMap}{SpatRast object containing depth information}

\item{AngleMap}{Optional SpatRast with D8 root angles (from terra::terrain(v="flowdir"))}

\item{RootMap}{Optional SpatRast containing segmented, presence-absence root image}

\item{select.layerRM}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `2`.}

\item{select.layerDM}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `NULL`.}

\item{select.layerAM}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `NULL`.}
}
\value{
If return_diagnostics=FALSE, returns numeric value representing the ratio of aligned root pixels
        to total root pixels. If TRUE, returns a list containing the ratio and diagnostic information.
}
\description{
This function analyzes root growth directions in relation to depth gradients.
It calculates the ratio of root pixels growing in the direction of the steepest
depth slope compared to all root pixels.
}
\examples{
data(skl_Oulanka2023_Session01_T067)
im = terra::rast(skl_Oulanka2023_Session01_T067)
DepthMap = terra::t(create.depthmap(im,center.offset=0,tube.thicc=3.5))

deep.drive(DepthMap = DepthMap, RootMap = im)
}
