% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats.R
\name{Directionality}
\alias{Directionality}
\title{Estimate the directional of a skeleton pixels}
\usage{
Directionality(im, rotate = TRUE)
}
\arguments{
\item{im}{skeletonized image raster. Roots must be 1, background 0. Consider the input image rotation to interpret the output}

\item{rotate}{should the image be rotated 90 degrees before entering?}
}
\value{
percentage of pixels with given neighbour pixel position
}
\description{
Estimate the directional of a skeleton pixels
}
\examples{
direction.frame = Directionality(im)
}
