% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats.R
\name{Directionality}
\alias{Directionality}
\title{Estimate the directional of a skeleton pixels}
\usage{
Directionality(im, rotate = TRUE)
}
\arguments{
\item{im}{a skeletonized image. Roots must be 1, background 0. The rotation of the image determines the output. The default is}
}
\value{
percentage of pixels with given neighbour pixel position
}
\description{
Estimate the directional of a skeleton pixels
}
\examples{
direction.frame = Directionality(im)
}
