% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats.R
\name{Halo}
\alias{Halo}
\title{Halo creates a buffer around pixel bigger than 0}
\usage{
Halo(im, width = 1, halo.only = T)
}
\arguments{
\item{im}{segmented raster}

\item{width}{buffer around roots in px}

\item{halo.only}{set TRUE if only the buffer around roots should be returned}
}
\value{
raster output
}
\description{
Halo creates a buffer around pixel bigger than 0
}
\examples{
buffIMG = Halo(im = im, width = 10, halo.only = T)
}
