% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateZoning3.R
\name{Halo}
\alias{Halo}
\title{Create a buffer halo) around non-zero pixels}
\usage{
Halo(img, width = 2, halo.only = TRUE, kernel = "circle")
}
\arguments{
\item{img}{SpatRaster/matrix/array - segmented image}

\item{width}{numeric - buffer width in pixels (default: 2)}

\item{halo.only}{logical - if TRUE, returns only the buffer zone (default: TRUE)}

\item{kernel}{character - shape of the thickening kernel: "circle" or "diamond"}
}
\value{
SpatRast - buffer zone around non-zero pixels
}
\description{
Create a buffer halo) around non-zero pixels
}
\examples{
data(seg_Oulanka2023_Session03_T067)
img <- terra::rast(seg_Oulanka2023_Session03_T067)
Halo(img, width = 2)
}
