% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lines_diagonal_botleft}
\alias{lines_diagonal_botleft}
\title{Example Image with circa right to left diagonal lines}
\format{
a RasterBrick with 194 columns x 100 rows x 4 layer
}
\source{
Images by J.Cunow
}
\usage{
data(lines_diagonal_botleft)
}
\description{
Use this dataset to test directionality
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
