% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lines_diagonal_topleft}
\alias{lines_diagonal_topleft}
\title{Example Image with Diagonal Lines (Top-Left to Bottom-Right)}
\format{
A RasterBrick object with dimensions:
  \itemize{
    \item 194 columns (width)
    \item 100 rows (height)
    \item 4 layers (channels)
  }
}
\source{
Images by J.Cunow
}
\usage{
data(lines_diagonal_topleft)
}
\description{
Test dataset containing an image with diagonal lines running from
top-left to bottom-right for directionality analysis.
}
\details{
The image consists of multiple diagonal lines suitable for testing
directionality detection algorithms.
}
\examples{
\donttest{
  data(lines_diagonal_topleft)
  plot(lines_diagonal_topleft)
}
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
