% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lines_horizontal}
\alias{lines_horizontal}
\title{Example Image with circa horizontal lines}
\format{
a RasterBrick with 184 columns x 90 rows x 4 layer
}
\source{
Images by J.Cunow
}
\usage{
data(lines_horizontal)
}
\description{
Use this dataset to test directionality
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
