% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lines_vertical}
\alias{lines_vertical}
\title{Example Image with circa vertical lines}
\format{
a RasterBrick with 144 columns x 99 rows x 4 layer
}
\source{
Images by J.Cunow
}
\usage{
data(lines_vertical)
}
\description{
Use this dataset to test directionality
}
\author{
Johannes Cunow \email{johannes.cunow@gmail.com}
}
\keyword{datasets}
