% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats.R
\name{px.sum}
\alias{px.sum}
\title{counts all pixels in a binarized root image}
\usage{
px.sum(root.zone)
}
\arguments{
\item{root.zone}{a one layer raster layer}
}
\value{
a numeric value
}
\description{
counts all pixels in a binarized root image
}
\examples{
value  = px.sum(root.zone)
}
