% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{rgb2gray}
\alias{rgb2gray}
\title{Converts RGB to Grayscale}
\usage{
rgb2gray(img, r = 0.21, g = 0.72, b = 0.07)
}
\arguments{
\item{img}{rgb raster}

\item{r}{weight for red color}

\item{g}{weight for green color}

\item{b}{weight for blue color}
}
\value{
a single layer gray scale raster
}
\description{
Converts RGB to Grayscale
}
\examples{
gray.raster = rgb2gray(img)
}
