% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{root.accumulation}
\alias{root.accumulation}
\title{Root accumulation Curve}
\usage{
root.accumulation(data, group, depth, variable)
}
\arguments{
\item{data}{dataframe must include group,depth, and variable columns}

\item{group}{specify the grouping variable e.g., Plot}

\item{depth}{specifiy column name which includes depth values}

\item{variable}{accumulating values}
}
\value{
dataframe with one added column "cs" containing the cummulated values
}
\description{
Root accumulation Curve
}
\examples{
data1 = root.accumulation(data,group = Plot, depth = depth, variable = rootpx)
}
