% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats_functions.R
\name{RootLength}
\alias{RootLength}
\title{Estimate Kimura Root Length}
\usage{
RootLength(im, unit = "cm", dpi = 300)
}
\arguments{
\item{im}{a skeletonized root image}

\item{unit}{output unit}

\item{dpi}{image resolution}
}
\value{
root length
}
\description{
Estimate Kimura Root Length
}
\examples{
RL = RootLength(im = image,unit = "cm", dpi = 300)
}
