% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats3.R
\name{RootLength}
\alias{RootLength}
\title{Calculate Root Length using Kimura's Method with optimizations}
\usage{
RootLength(img, unit = "cm", dpi = 300, select.layer = NULL)
}
\arguments{
\item{img}{A skeletonized root image raster}

\item{unit}{Output unit ("px" or "cm")}

\item{dpi}{Image resolution (required when unit = "cm")}

\item{select.layer}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `2`.}
}
\value{
Numeric value representing root length in specified unit
}
\description{
Calculate Root Length using Kimura's Method with optimizations
}
\examples{
data(skl_Oulanka2023_Session01_T067)
img = terra::rast(skl_Oulanka2023_Session01_T067)
RL = RootLength(img = img,unit = "cm", dpi = 300, select.layer = 2)
}
