% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_kimurarootlength_function.R
\name{RootLength}
\alias{RootLength}
\title{Estimate Kimura Root Length}
\usage{
RootLength(im, unit = "cm", dpi = 300)
}
\arguments{
\item{im}{a skeletonized root image}

\item{unit}{output unit}

\item{dpi}{image resolution}
}
\value{
numeric value of root length
}
\description{
Estimate Kimura Root Length
}
\examples{
value = RootLength(skeletonized.image)
}
