% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats.R
\name{RootLength}
\alias{RootLength}
\title{Estimate Kimura Root Length}
\usage{
RootLength(im, unit = "cm", dpi = 300)
}
\arguments{
\item{im}{a skeletonized root image raster}

\item{unit}{output unit}

\item{dpi}{image resolution}
}
\value{
root length
}
\description{
Estimate Kimura Root Length
}
\examples{
data(skl_Oulanka2023_Session01_T067)
img = terra::rast(skl_Oulanka2023_Session01_T067[[2]])
RL = RootLength(im = img,unit = "cm", dpi = 300)
}
