% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_RootScape.R
\name{RootScapeMetrics}
\alias{RootScapeMetrics}
\title{RootScape relies on Landscapemetrics as working horse}
\usage{
RootScapeMetrics(
  im,
  indexD,
  metrics = c("lsm_c_ca", "lsm_l_ent", "lsm_c_pd", "lsm_c_np", "lsm_c_pland",
    "lsm_c_area_mn", "lsm_c_area_cv", "lsm_c_enn_mn", "lsm_c_enn_cv")
)
}
\arguments{
\item{im}{segmented, unskeletonized raster}

\item{metrics}{which ,metrics should be calculated from the available ones in 'landscapemetrics::calculate_lsm()'. A selection is used as default
Interpretation here: ...}
}
\value{
a bunch of metric values
}
\description{
RootScape relies on Landscapemetrics as working horse
}
\examples{
RootScapeObject  = RootScapeMetrics(image)
}
