% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Rotational_function.R
\name{RotationE}
\alias{RotationE}
\title{Estimates Rotation from Tape Coverage}
\usage{
RotationE(img, tape.brightness = 0.95)
}
\arguments{
\item{img}{image.tiff}

\item{tape.brightness}{used for clustering. Tape appears bright e.g., 0.95}
}
\value{
numeric value corresponding to the center of extruding tape
}
\description{
Estimates Rotation from Tape Coverage
}
\examples{
value = RotationE(img)
}
