% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{RotCensor}
\alias{RotCensor}
\title{Censor image edges based on rotation}
\usage{
RotCensor(
  img,
  center.offset = 0,
  cut.buffer = 0.02,
  fixed.rotation = TRUE,
  fixed.width = 1000,
  select.layer = NULL
)
}
\arguments{
\item{img}{Input image to censor}

\item{center.offset}{Rotation shift in rows (from RotShiftDet)}

\item{cut.buffer}{Proportion of image to cut when fixed_rotation=FALSE}

\item{fixed.rotation}{Use fixed output dimensions}

\item{fixed.width}{Output width when fixed_rotation=TRUE}

\item{select.layer}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `NULL`.}
}
\value{
Cropped raster image
}
\description{
Crops image edges to handle non-overlapping regions between sequential scans.
}
\examples{
data(seg_Oulanka2023_Session01_T067)
img = terra::rast(seg_Oulanka2023_Session01_T067)
censored.raster = RotCensor(img,center.offset = 120, cut.buffer = 0.02)
}
