% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{RotShiftDet}
\alias{RotShiftDet}
\title{Rotation Correlation}
\usage{
RotShiftDet(img1, img2, cor.type = "phase")
}
\arguments{
\item{img1}{Reference Image}

\item{img2}{Subsequent Image}

\item{cor.type}{Two correlation types available: "ccf" cross correlation, and "phase" phase correlation in frequency domain. See ??imagefx}
}
\value{
numeric value corresponding to rotation in rows. Ensure correct image rotation i.e., rows == rotation
}
\description{
Rotation Correlation
}
\examples{
y.lag = RotShiftDet(img,img2,"phase")
}
