% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{RotShiftDet}
\alias{RotShiftDet}
\title{Detect rotation shift between two images}
\usage{
RotShiftDet(
  img1,
  img2,
  cor.type = "phase",
  fixed.depth.pixel = c(1000, 4000),
  fixed.width = NULL,
  select.layer = NULL
)
}
\arguments{
\item{img1}{Reference image (3-channel RGB)}

\item{img2}{Subsequent image to compare (3-channel RGB)}

\item{cor.type}{Correlation type: "ccf" (cross) or "phase" (frequency domain)}

\item{fixed.depth.pixel}{Depth range to analyze c(start, end)}

\item{fixed.width}{Width of analysis region in pixels}

\item{select.layer}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `NULL`.}
}
\value{
Vector of shifts (x,y) in pixels
}
\description{
Calculates the rotation shift between two sequential images using either
cross-correlation or phase correlation methods.
}
\examples{
img1 = seg_Oulanka2023_Session01_T067
img2 = seg_Oulanka2023_Session03_T067
y.lag = RotShiftDet(img1,img2,"phase")
}
