% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDistributions.R
\name{RPI}
\alias{RPI}
\title{Root Penetration Index}
\usage{
RPI(roots, w)
}
\arguments{
\item{roots}{root cover}

\item{w}{weight, typically a vector containing depths}
}
\value{
a value between 0-1
}
\description{
Root Penetration Index
}
\examples{
w = seq(5,25,5)
roots = c(0,10,7,3,1)
RPI(w,roots)
}
