% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LilHelpers.R
\name{skeletonize}
\alias{skeletonize}
\title{Skeletonize a segmented image}
\usage{
skeletonize(img, itr = 2, kernel = "Skeleton:3")
}
\arguments{
\item{img}{can be: path, array, raster, or magick object}

\item{itr}{iterations of thinning. see magick::image_morphology}

\item{kernel}{kernel used for image operation. Other operations than skeletonizing is possible.max(se)}
}
\value{
image in as magick object
}
\description{
Skeletonize a segmented image
}
\details{
Uses magick::image_morphology but eases the image input format
}
\examples{
img.skeleton = skeletonize(skl_Oulanka2023_Session01_T067,itr = 2)
}
