% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_SoilSurfaceEstimation_function.R
\name{SoilSurfE}
\alias{SoilSurfE}
\title{Estimate the position of the soil surface by tape presence}
\usage{
SoilSurfE(
  im,
  search.area = 0.45,
  tape.tresh = 0.33,
  dpi = 300,
  tape.overlap = 0.5,
  tape.brightness = 0.66,
  extra.rows = 100
)
}
\arguments{
\item{im}{image.tiff}

\item{search.area}{ratio of image which is used to look for tape cover. Speeds up computation.}

\item{tape.tresh}{ratio of how much of the tube rotation needs to covered in tape}

\item{dpi}{image resolution}

\item{tape.overlap}{assumes a saftey margin on the tape. The soil surface will be shifted by this amount in cm}

\item{tape.brightness}{used for clustering. Tape appears bright e.g., 0.95}

\item{extra.rows}{In case no tape is present. Best leave unchanged - some extra.rows are recommended and will be substracted from the output anyway.}
}
\value{
data.frame with tape end and soil surface estimation in rows
}
\description{
Estimate the position of the soil surface by tape presence
}
\examples{
data.frame(soil0, tape.end) = SoilSurfE(im)
}
