% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration3.R
\name{SoilSurfE}
\alias{SoilSurfE}
\title{Estimate soil surface position using tape markers}
\usage{
SoilSurfE(
  img,
  search.area = 0.45,
  tape.tresh = 0.33,
  dpi = 150,
  nclasses = 3,
  inverse = FALSE,
  tape.overlap = 0.5,
  tape.brightness = 0.6,
  extra.rows = 100,
  tape.quantile = 0.98,
  select.layer = NULL
)
}
\arguments{
\item{img}{Input image (raster, filename, or array)}

\item{search.area}{Proportion of image to analyze}

\item{tape.tresh}{Minimum tape coverage ratio}

\item{dpi}{Image resolution}

\item{nclasses}{Number of clustering classes}

\item{inverse}{Invert detection for dark markers}

\item{tape.overlap}{Safety margin for tape (cm)}

\item{tape.brightness}{Brightness threshold for tape}

\item{extra.rows}{Additional analysis rows}

\item{tape.quantile}{Brightness alignment quantile}

\item{select.layer}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `NULL`.}
}
\value{
data.frame with soil surface and tape end positions
}
\description{
Detects the soil surface by analyzing tape coverage patterns in the image.
}
\examples{
img = rgb_Oulanka2023_Session03_T067
Soil0Estimates = SoilSurfE(img)
}
