% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats.R
\name{texture}
\alias{texture}
\title{Enhanced texture calculation}
\usage{
texture(
  img.color,
  grays = 7,
  window = c(9, 9),
  metrics = c("variance", "second_moment")
)
}
\arguments{
\item{img.color}{Three-band raster or path to image}

\item{grays}{Number of gray levels}

\item{window}{Window size for GLCM}

\item{metrics}{Texture metrics to calculate}
}
\value{
Raster with texture metrics
}
\description{
Enhanced texture calculation
}
\examples{
data(rgb_Oulanka2023_Session03_T067)
img = raster::brick(rgb_Oulanka2023_Session03_T067)
texture(img, 7, c(9,9), metrics = "second_moment")
}
