% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats.R
\name{texture}
\alias{texture}
\title{Texture corresponds to glmc::glmc()}
\usage{
texture(
  img.color,
  grays = 7,
  window = c(9, 9),
  metrics = c("variance", "second_moment")
)
}
\arguments{
\item{img.color}{image with three color channels. Will be converted in gray scale. raster format required.}

\item{grays}{number of gray shades. Documentation is lacking, see: ?glmc::glmc()}

\item{window}{convolution window size e.g., c(3,3)}

\item{metrics}{texture metrics based on illumination differences. see:: ?glmc::glmc() for available methods}
}
\value{
raster with each layer correspond to on metric
}
\description{
Texture corresponds to glmc::glmc()
}
\examples{
data(rgb_Oulanka2023_Session03_T067)
img = rgb_Oulanka2023_Session03_T067
texture(img, 7, c(9,9), metrics = "second_moment")
}
