% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats.R
\name{Tube.coloration}
\alias{Tube.coloration}
\title{Calculate Image Coloration Metrics}
\usage{
Tube.coloration(img, r = 0.2126, g = 0.7152, b = 0.0722)
}
\arguments{
\item{img}{Three-band raster (RGB) or path to image}

\item{r}{Red channel weight}

\item{g}{Green channel weight}

\item{b}{Blue channel weight}
}
\value{
Data frame of color metrics
}
\description{
Calculate Image Coloration Metrics
}
\examples{
data(rgb_Oulanka2023_Session03_T067)
img = terra::rast(rgb_Oulanka2023_Session03_T067)
colorvector = Tube.coloration(img)
}
