% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zonestats_functions.R
\name{Tube.coloration}
\alias{Tube.coloration}
\title{Coloration of the image}
\usage{
Tube.coloration(img, r = 0.2126, g = 0.7152, b = 0.0722)
}
\arguments{
\item{img}{raster with 3 color bands}

\item{r}{weight for first channel - typically red}

\item{g}{weight for second channel - typically green}

\item{b}{weight for third channel - typically blue}
}
\value{
a vector with chromatic coordinates,luminosity, brightness, luminosity, color values, saturation
}
\description{
Coloration of the image
}
\examples{
colorvector = Tube.coloration(img = imrgb)
}
