% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turnover.R
\name{Turnover.PDC}
\alias{Turnover.PDC}
\title{Estimates New Root Production, Root Decay, and Roots without change}
\usage{
Turnover.PDC(
  img,
  product.layer = 2,
  decay.layer = 1,
  blur.capture = 0.95,
  im.return = FALSE,
  include.virtualroots = TRUE
)
}
\arguments{
\item{img}{image in the 'RootDetector' format}

\item{product.layer}{layer indicating production}

\item{decay.layer}{layer indicating decay & tape}

\item{blur.capture}{saftey margin on production-decay-constant delination}

\item{im.return}{return images instead of values?}

\item{include.virtualroots}{should all roots which were present at some point in one of the two time steps be considered?}
}
\value{
either pixel sums if im.return = F, or individual layers corresponding to tape, production, decay, and no change
}
\description{
Estimates New Root Production, Root Decay, and Roots without change
}
\examples{
 PDCs = Turnover.PDC(img = img, im.return = F)
}
