% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateZoning3.R
\name{zone.fun}
\alias{zone.fun}
\title{Extract a zone of interest based on depth values}
\usage{
zone.fun(rootpic, binned.map, indexD = 0, silent = FALSE, select.layer = NULL)
}
\arguments{
\item{rootpic}{SpatRaster/matrix/array - source image}

\item{binned.map}{SpatRaster/matrix/array - binned depth map}

\item{indexD}{numeric - depth index to extract. Usually used to index in a loop to iterate over depths.}

\item{silent}{logical - suppress messages}

\item{select.layer}{Integer. Specifies which layer to use if the input is a multi-band image. Default is `2`.}
}
\value{
SpatRaster - extracted zone
}
\description{
Extract a zone of interest based on depth values
}
\examples{
data(seg_Oulanka2023_Session01_T067)
img = terra::rast(seg_Oulanka2023_Session01_T067)
mask = img[[1]] - img[[2]]
mask[mask == 255] <- NA
img = img
depthmap = create_depthmap(img,mask,start.soil = 2.9, select.layer = 2 )
binned.map = binning(depthmap,nn = 5)

data(seg_Oulanka2023_Session01_T067)
img = terra::rast(seg_Oulanka2023_Session01_T067)
mask = img[[1]] - img[[2]]
mask[mask == 255] <- NA
depthmap = create_depthmap(img,mask,start.soil = 0, tube.thicc = 7,dpi = 150, select.layer = 2 )
binned.map = binning(depthmap,nn = 5)
image.zone = zone.fun(img, binned.map, indexD = 10,  silent = FALSE,select.layer = NULL)
}
