% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateZoning.R
\name{zone.fun}
\alias{zone.fun}
\title{Cuts out a zone of interest}
\usage{
zone.fun(rootpic, binned.map, indexD = 0, nn = 5, silent = FALSE)
}
\arguments{
\item{rootpic}{the image which should be cut}

\item{binned.map}{image which supplies the cut condition}

\item{indexD}{the condition}

\item{nn}{bin width in binned.map. Should be the same as used in 'binning()'}

\item{silent}{verbose}
}
\value{
a cutout of the root pic image
}
\description{
Cuts out a zone of interest
}
\examples{
img = seg_Oulanka2023_Session01_T067[[2]]
mask = seg_Oulanka2023_Session01_T067[[1]] - seg_Oulanka2023_Session01_T067[[2]]
mask[mask == 255] <- NA
depthmap = create.depthmap(img,mask,start.soil = 290 )
binned.map = binning(depthmap,nn = 5)
image.zone = zone.fun(img, binned.map, indexD = 0, nn=5, silent = FALSE)
}
