% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Helper_functions.R
\name{zone.fun}
\alias{zone.fun}
\title{Cuts out a zone of interest}
\usage{
zone.fun(rootpic, binned.map, indexD, nn = 5, silent = F)
}
\arguments{
\item{rootpic}{the image which should be cut}

\item{binned.map}{image which supplies the cut condition}

\item{indexD}{the condition}

\item{nn}{bin width in binned.map. Should be the same as used in 'binning()'}

\item{silent}{verbose}
}
\value{
a cutout of the root pic image
}
\description{
Cuts out a zone of interest
}
\examples{
raster_image = zone.fun(rootpic, binned.map, nn=5, silent = F)
}
