% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zoning_functions.R
\name{zone.rotation.fun}
\alias{zone.rotation.fun}
\title{RotationZones}
\usage{
zone.rotation.fun(rootpic, k = c(3, 4), nn = 5, mm = c(2000, 5000))
}
\arguments{
\item{rootpic}{the "to be cut" image}

\item{k}{specifying which divisions to keep. Must be <= nn}

\item{nn}{number of total cut divisions}

\item{mm}{region along the tube. Adjust to your tube dimensions!}
}
\value{
raster image
}
\description{
RotationZones
}
\examples{
rotationZone1 = zone.rotation.fun(rootpic, k = c(1,2), nn = 7, mm = c(1500,3000))
}
